+++ Gemfile
@@ -1,4 +1,4 @@
 source 'http://rubygems.org'
 
 gemspec
-
+gem "pry"
