+++ lib/common/version.rb
@@ -1,0 +1,3 @@
+module Rubyplot
+  VERSION = '0.1'.freeze
+end
