--- lib/grruby/grruby.rb
@@ -1,11 +1,0 @@
-# GR Ruby C Extension Files
-require 'grruby.so'
-require 'grruby/tasks/gr_function_classes'
-
-# Scripting Layer files of GR Plot.
-require 'grruby/scripting/plots'
-require 'grruby/scripting/plot/bar'
-require 'grruby/scripting/plot/scatter'
-require 'grruby/scripting/plot/stacked_bar'
-require 'grruby/scripting/plotspace'
-require 'grruby/scripting/subplot'
