--- lib/grruby/scatter.rb
@@ -1,9 +1,0 @@
-module Rubyplot
-  module GR
-    class Scatter
-      def data x_data=[], y_data=[], label: :default, color: :default
-        
-      end
-    end
-  end
-end
