--- lib/rubyplot/figure/geometry.rb
@@ -1,9 +1,0 @@
-module Rubyplot
-  class Figure
-    # TODO: extract instance variables from Figure into this class. Maybe rename.
-    class Geometry
-      def initialize 
-      end
-    end
-  end
-end
