--- lib/rubyplot/version.rb
@@ -1,3 +1,0 @@
-module Rubyplot
-  VERSION = '0.1'.freeze
-end
