+++ rubyplot.gemspec
@@ -13,14 +13,12 @@ Gem::Specification.new do |s|
   s.files       = `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(test|spec|features)/}) }
   s.license     = 'MIT license'
   s.homepage    = 'https://github.com/Arafatk/magick-rubyplot'
-  s.extensions  = ['ext/grruby/extconf.rb']
-  
   s.add_development_dependency 'bundler'
   s.add_development_dependency 'rake'
-  s.add_development_dependency 'rake-compiler'
   s.add_dependency 'rmagick',  '>= 2.13.4'
   s.add_development_dependency 'gtk3'
   s.add_development_dependency 'rspec'
   s.add_development_dependency 'rspec-autotest'
   s.add_development_dependency 'ZenTest'
+  s.extensions = ['ext/grruby/extconf.rb']
 end
