+++ spec/bar/bar_spec.rb
@@ -1,5 +1,5 @@
 require 'spec_helper'
-RubyplotSpec.describe 'Bar' do
+describe 'Bar' do
   it 'Sets up a basic reference image for bar graph with random numbers' do
     random_bar_graph
     plot = Rubyplot::Bar.new(600)
