+++ spec/spec_helper.rb
@@ -10,21 +10,6 @@ require 'scatter/scatter_gen'
 require 'bubble/bubble_gen'
 require 'stacked_bar/stacked_gen'
 
-class RubyplotSpec
-  def self.describe *args, &block
-    [:GR, :magick].each do |backend|
-      Rubyplot.backend = backend
-      ::RSpec.describe args, &block
-    end
-  end
-end
-
-RSpec::Matchers.define :eq_image do |expected_image, delta|
-  match do |actual_image|
-    compare_with_reference? actual_image, expected_image, delta
-  end
-end
-
 def compare_with_reference?(test_image, reference_image, tolerance)
   compute_rms(test_image, reference_image) < tolerance
 end
@@ -33,8 +18,6 @@ end
 def compute_rms(test_image, reference_image)
   image1 = Magick::Image.read(('spec/reference_images/' + test_image)).first
   image2 = Magick::Image.read(('spec/reference_images/' + reference_image)).first
-  puts "image1: #{image1.export_pixels.size}"
-  puts "image2: #{image2.export_pixels.size}"
   diff = 0
   pixel_array_1 = image1.export_pixels
   pixel_array_2 = image2.export_pixels
